refactor; PostManagerPage의 Dialog 컴포넌트 분리
