+++ src/entities/postsManager/PostsManagerPage.tsx
@@ -1,63 +1,31 @@
 import { useEffect, useState } from "react"
-import { Plus, Search } from "lucide-react"
-import {
-  Button,
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  Input,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-  Textarea,
-  useQueryParams,
-} from "../../shared"
-import { HighlightText } from "@/shared"
-import { Comments } from "../comments/ui/comments"
+import { Plus } from "lucide-react"
+import { Button, Card, CardContent, CardHeader, CardTitle, useQueryParams } from "@/shared"
 import { PostsTable } from "../posts/ui/PostsTable"
 import {
-  addPost,
   getCommentsByPostId,
   getPosts,
   getPostsBySearch,
   getPostsByTag,
-  getTags,
   getUsers,
   Post,
-  Tag,
-  updatePost,
   User,
   Comment,
   getUserById,
   UserInfo,
 } from "./apis/postsApis"
+import { SearchFilterContainer } from "./SearchFilterContainer"
+import { Paginations } from "./Paginations"
+import { AddPostDialog } from "./AddPostDialog"
+import { EditPostDialog } from "./apis/EditPostDialog"
+import { PostDetailDialog } from "./PostDetailDialog"
+import { UserDialog } from "./UserDialog"
 
-interface PostWithAuthor extends Post {
+export interface PostWithAuthor extends Post {
   author?: User
 }
 export const PostsManagerComponent = () => {
-  const {
-    skip,
-    limit,
-    searchQuery,
-    sortBy,
-    sortOrder,
-    selectedTag,
-    updateURL,
-    setSkip,
-    setLimit,
-    setSearchQuery,
-    setSortBy,
-    setSortOrder,
-    setSelectedTag,
-  } = useQueryParams()
+  const { skip, limit, searchQuery, sortBy, sortOrder, selectedTag, updateURL, setSelectedTag } = useQueryParams()
 
   // 상태 관리
   const [posts, setPosts] = useState<PostWithAuthor[]>([])
@@ -66,9 +34,8 @@ export const PostsManagerComponent = () => {
   const [selectedPost, setSelectedPost] = useState<PostWithAuthor>()
   const [showAddDialog, setShowAddDialog] = useState(false)
   const [showEditDialog, setShowEditDialog] = useState(false)
-  const [newPost, setNewPost] = useState({ title: "", body: "", userId: 1 })
   const [loading, setLoading] = useState(false)
-  const [tags, setTags] = useState<Tag[]>([])
+
   const [comments, setComments] = useState<Record<number, Comment[]>>({})
   const [showPostDetailDialog, setShowPostDetailDialog] = useState(false)
   const [showUserModal, setShowUserModal] = useState(false)
@@ -87,16 +54,6 @@ export const PostsManagerComponent = () => {
     setPosts(postsWithUsers)
     setTotal(postsData.total)
     setLoading(false)
-  }
-
-  // 태그 가져오기
-  const fetchTags = async () => {
-    try {
-      const tagsData = await getTags()
-      setTags(tagsData)
-    } catch (error) {
-      console.error("태그 가져오기 오류:", error)
-    }
   }
 
   // 게시물 검색
@@ -148,46 +105,6 @@ export const PostsManagerComponent = () => {
     setLoading(false)
   }
 
-  // 게시물 추가
-  const handleAddPost = async () => {
-    try {
-      const newPostData = await addPost(newPost)
-      setPosts([newPostData, ...posts])
-      setShowAddDialog(false)
-      //초기화
-      setNewPost({ title: "", body: "", userId: 1 })
-    } catch (error) {
-      console.error("게시물 추가 오류:", error)
-    }
-  }
-
-  const openEditDialog = (post: PostWithAuthor) => {
-    setSelectedPost(post)
-    setShowEditDialog(true)
-  }
-
-  // 게시물 업데이트
-  const handleUpdatePost = async () => {
-    if (!selectedPost) return
-
-    try {
-      const editedPost: Post = {
-        id: selectedPost.id,
-        title: selectedPost.title,
-        body: selectedPost.body,
-        userId: selectedPost.userId,
-        tags: selectedPost.tags,
-        reactions: selectedPost.reactions,
-        views: selectedPost.views,
-      }
-      const updatedPost = await updatePost(editedPost)
-      setPosts(posts.map((post) => (post.id === updatedPost.id ? updatedPost : post)))
-      setShowEditDialog(false)
-    } catch (error) {
-      console.error("게시물 업데이트 오류:", error)
-    }
-  }
-
   // 댓글 가져오기
   const fetchComments = async (postId: number) => {
     if (comments[postId]) return // 이미 불러온 댓글이 있으면 다시 불러오지 않음
@@ -197,6 +114,12 @@ export const PostsManagerComponent = () => {
     } catch (error) {
       console.error("댓글 가져오기 오류:", error)
     }
+  }
+
+  // 게시물 수정 모달 열기
+  const openEditDialog = (post: PostWithAuthor) => {
+    setSelectedPost(post)
+    setShowEditDialog(true)
   }
 
   // 게시물 상세 보기
@@ -216,10 +139,6 @@ export const PostsManagerComponent = () => {
       console.error("사용자 정보 가져오기 오류:", error)
     }
   }
-
-  useEffect(() => {
-    fetchTags()
-  }, [])
 
   useEffect(() => {
     if (selectedTag) {
@@ -244,61 +163,12 @@ export const PostsManagerComponent = () => {
       <CardContent>
         <div className="flex flex-col gap-4">
           {/* 검색 및 필터 컨트롤 */}
-          <div className="flex gap-4">
-            <div className="flex-1">
-              <div className="relative">
-                <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
-                <Input
-                  placeholder="게시물 검색..."
-                  className="pl-8"
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  onKeyPress={(e) => e.key === "Enter" && searchPosts()}
-                />
-              </div>
-            </div>
-            <Select
-              value={selectedTag}
-              onValueChange={(value) => {
-                setSelectedTag(value)
-                fetchPostsByTag(value)
-                updateURL()
-              }}
-            >
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="태그 선택" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">모든 태그</SelectItem>
-                {tags.map((tag) => (
-                  <SelectItem key={tag.url} value={tag.slug}>
-                    {tag.slug}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-            <Select value={sortBy} onValueChange={setSortBy}>
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="정렬 기준" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="none">없음</SelectItem>
-                <SelectItem value="id">ID</SelectItem>
-                <SelectItem value="title">제목</SelectItem>
-                <SelectItem value="reactions">반응</SelectItem>
-              </SelectContent>
-            </Select>
-            <Select value={sortOrder} onValueChange={setSortOrder}>
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="정렬 순서" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="asc">오름차순</SelectItem>
-                <SelectItem value="desc">내림차순</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
-
+          <SearchFilterContainer
+            searchPosts={searchPosts}
+            selectedTag={selectedTag}
+            setSelectedTag={setSelectedTag}
+            fetchPostsByTag={fetchPostsByTag}
+          />
           {/* 게시물 테이블 */}
           {loading ? (
             <div className="flex justify-center p-4">로딩 중...</div>
@@ -313,135 +183,42 @@ export const PostsManagerComponent = () => {
           )}
 
           {/* 페이지네이션 */}
-          <div className="flex justify-between items-center">
-            <div className="flex items-center gap-2">
-              <span>표시</span>
-              <Select value={limit.toString()} onValueChange={(value) => setLimit(Number(value))}>
-                <SelectTrigger className="w-[180px]">
-                  <SelectValue placeholder="10" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="10">10</SelectItem>
-                  <SelectItem value="20">20</SelectItem>
-                  <SelectItem value="30">30</SelectItem>
-                </SelectContent>
-              </Select>
-              <span>항목</span>
-            </div>
-            <div className="flex gap-2">
-              <Button disabled={skip === 0} onClick={() => setSkip(Math.max(0, skip - limit))}>
-                이전
-              </Button>
-              <Button disabled={skip + limit >= total} onClick={() => setSkip(skip + limit)}>
-                다음
-              </Button>
-            </div>
-          </div>
+          <Paginations total={total} />
         </div>
       </CardContent>
 
       {/* 게시물 추가 대화상자 */}
-      <Dialog open={showAddDialog} onOpenChange={setShowAddDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>새 게시물 추가</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Input
-              placeholder="제목"
-              value={newPost.title}
-              onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
-            />
-            <Textarea
-              rows={30}
-              placeholder="내용"
-              value={newPost.body}
-              onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
-            />
-            <Input
-              type="number"
-              placeholder="사용자 ID"
-              value={newPost.userId}
-              onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
-            />
-            <Button onClick={handleAddPost}>게시물 추가</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
-
+      <AddPostDialog
+        showAddDialog={showAddDialog}
+        setShowAddDialog={setShowAddDialog}
+        posts={posts}
+        setPosts={setPosts}
+      />
       {/* 게시물 수정 대화상자 */}
-      <Dialog open={showEditDialog} onOpenChange={setShowEditDialog}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>게시물 수정</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Input
-              placeholder="제목"
-              value={selectedPost?.title || ""}
-              onChange={(e) => setSelectedPost({ ...selectedPost, title: e.target.value } as PostWithAuthor)}
-            />
-            <Textarea
-              rows={15}
-              placeholder="내용"
-              value={selectedPost?.body || ""}
-              onChange={(e) => setSelectedPost({ ...selectedPost, body: e.target.value } as PostWithAuthor)}
-            />
-            <Button onClick={handleUpdatePost}>게시물 업데이트</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
+      {selectedPost && (
+        <EditPostDialog
+          showEditDialog={showEditDialog}
+          setShowEditDialog={setShowEditDialog}
+          selectedPost={selectedPost}
+          setSelectedPost={setSelectedPost}
+          posts={posts}
+          setPosts={setPosts}
+        />
+      )}
 
       {/* 게시물 상세 보기 대화상자 */}
-      <Dialog open={showPostDetailDialog} onOpenChange={setShowPostDetailDialog}>
-        <DialogContent className="max-w-3xl">
-          <DialogHeader>
-            <DialogTitle>
-              <HighlightText text={selectedPost?.title || ""} highlight={searchQuery} />
-            </DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <p>
-              <HighlightText text={selectedPost?.body || ""} highlight={searchQuery} />
-            </p>
-            {selectedPost && <Comments postId={selectedPost.id} />}
-          </div>
-        </DialogContent>
-      </Dialog>
+      {selectedPost && (
+        <PostDetailDialog
+          showPostDetailDialog={showPostDetailDialog}
+          setShowPostDetailDialog={setShowPostDetailDialog}
+          selectedPost={selectedPost}
+        />
+      )}
 
       {/* 사용자 모달 */}
-      <Dialog open={showUserModal} onOpenChange={setShowUserModal}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>사용자 정보</DialogTitle>
-          </DialogHeader>
-          <div className="space-y-4">
-            <img src={selectedUser?.image} alt={selectedUser?.username} className="w-24 h-24 rounded-full mx-auto" />
-            <h3 className="text-xl font-semibold text-center">{selectedUser?.username}</h3>
-            <div className="space-y-2">
-              <p>
-                <strong>이름:</strong> {selectedUser?.firstName} {selectedUser?.lastName}
-              </p>
-              <p>
-                <strong>나이:</strong> {selectedUser?.age}
-              </p>
-              <p>
-                <strong>이메일:</strong> {selectedUser?.email}
-              </p>
-              <p>
-                <strong>전화번호:</strong> {selectedUser?.phone}
-              </p>
-              <p>
-                <strong>주소:</strong> {selectedUser?.address?.address}, {selectedUser?.address?.city},{" "}
-                {selectedUser?.address?.state}
-              </p>
-              <p>
-                <strong>직장:</strong> {selectedUser?.company?.name} - {selectedUser?.company?.title}
-              </p>
-            </div>
-          </div>
-        </DialogContent>
-      </Dialog>
+      {selectedUser && (
+        <UserDialog showUserModal={showUserModal} setShowUserModal={setShowUserModal} selectedUser={selectedUser} />
+      )}
     </Card>
   )
 }
